Progressing from what was built.
I added in enemies and an enemy spawner along with road which will track the enemies movements within the map, various upgrades were made to the background and the GUI to make it fit more to the theme. The player turret was now in the game allowing the player to place them, addition to this the player can also upgrade their turret to all 3 levels.